+++ test/lib/index.spec.js
@@ -235,95 +235,6 @@ describe('The API', function() {
 
 describe('Mocking the file system', function() {
 
-  describe('fs.writeFile(filename, data, [options], callback)', function() {
-    beforeEach(function() {
-      mock({
-        dir: mock.directory({
-          mtime: new Date(1)
-        })
-      });
-    });
-    afterEach(mock.restore);
-
-    it('writes a string to a file', function(done) {
-      fs.writeFile('dir/foo', 'bar', function(err) {
-        if (err) {
-          return done(err);
-        }
-        assert.equal(String(fs.readFileSync('dir/foo')), 'bar');
-        done();
-      });
-    });
-
-    if (fs.promises) {
-      it('writes a string to a file in promise', function(done) {
-        fs.promises.writeFile('dir/foo', 'bar').then(
-          function() {
-            assert.equal(String(fs.readFileSync('dir/foo')), 'bar');
-            done();
-          },
-          function(err) {
-            done(err);
-          }
-        );
-      });
-    }
-
-    it('updates mtime of parent directory', function(done) {
-      const oldTime = fs.statSync('dir').mtime;
-      fs.writeFile('dir/foo', 'bar', function(err) {
-        if (err) {
-          return done(err);
-        }
-        const newTime = fs.statSync('dir').mtime;
-        assert.isTrue(newTime > oldTime);
-        done();
-      });
-    });
-
-    it('writes a buffer to a file', function(done) {
-      fs.writeFile('dir/foo', bufferFrom('bar'), function(err) {
-        if (err) {
-          return done(err);
-        }
-        assert.equal(String(fs.readFileSync('dir/foo')), 'bar');
-        done();
-      });
-    });
-
-    it('fails if directory does not exist', function(done) {
-      fs.writeFile('foo/bar', 'baz', function(err) {
-        assert.instanceOf(err, Error);
-        done();
-      });
-    });
-  });
-
-  describe('fs.writeFileSync(filename, data, [options]', function() {
-    beforeEach(function() {
-      mock({
-        '.': {}
-      });
-    });
-    afterEach(mock.restore);
-
-    it('writes a string to a file', function() {
-      fs.writeFileSync('foo', 'bar');
-      assert.equal(String(fs.readFileSync('foo')), 'bar');
-    });
-
-    it('writes a buffer to a file', function() {
-      fs.writeFileSync('foo', bufferFrom('bar'));
-      assert.equal(String(fs.readFileSync('foo')), 'bar');
-    });
-
-    it('fails if directory does not exist', function() {
-      assert.throws(function() {
-        fs.writeFileSync('foo/bar', 'baz');
-      });
-    });
-  });
-
   describe('fs.appendFile(filename, data, [options], callback)', function() {
     beforeEach(function() {
       mock({
